kobrasuitecore/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cors==1.0.1
cryptography==44.0.0
daphne==4.1.2
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
dos2unix==1
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
factory_boy==3.3.1
Faker==33.3.0
filelock==3.16.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
msgpack==1.1.0
openai==1.59.6
packaging==24.2
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==24.3.0
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.1
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.1
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
rpds-py==0.22.3
sentry-sdk==2.20.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tldextract==5.1.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==14.1
whitenoise==6.6.0
zope.event==5.0
zope.interface==7.2
